# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Shortens URLs and resolves short codes back to their original URL.",
        "title": "URL Shortener API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/links": {
            "get": {
                "description": "Returns every stored short code and its original URL.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "links"
                ],
                "summary": "List all links",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.listResponse"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/shorten": {
            "post": {
                "description": "Creates a short code for the given URL.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "links"
                ],
                "summary": "Shorten a URL",
                "parameters": [
                    {
                        "description": "URL to shorten",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.shortenRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.shortenResponse"
                        }
                    },
                    "400": {
                        "description": "invalid request or URL",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/{code}": {
            "get": {
                "description": "Redirects to the original URL for the given code.",
                "tags": [
                    "links"
                ],
                "summary": "Resolve a short code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Short code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found"
                    },
                    "400": {
                        "description": "invalid short url",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "short url not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.listResponse": {
            "type": "object",
            "properties": {
                "links": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/shortener.Link"
                    }
                }
            }
        },
        "handler.shortenRequest": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        },
        "handler.shortenResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                }
            }
        },
        "shortener.Link": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "originalURL": {
                    "type": "string"
                }
            }
        }
    }
}